Backend/requirements.txt
fastapi
uvicorn
pydantic
chromadb
pymongo
google-generativeai
sentence-transformers
python-dotenv
semantic-kernel
